requirements.txt
# Data Collection & Analysis
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Report Generation
jinja2>=3.1.0
weasyprint>=60.0
reportlab>=4.0.0

# Dashboard
streamlit>=1.28.0

# LLM Integration (Optional - for AI insights)
openai>=1.0.0
anthropic>=0.8.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0
